EF-3 - Popup próprio funcionando e sombras adicionadas

+++ erp/lib/templates/pagina_com_appbar/pagina_com_appbar_controller.dart
@@ -18,7 +18,7 @@ class PaginaComAppBarController extends GetxController {
     _modoNoturno.value = !modoNoturno;
   }
 
-  void exibirMenu(double value) {
+  void exibirMenu(int value) {
     menuExibido = (value != _menuExibido.value) ? value : double.infinity;
   }
